+++ shims/spark35/src/main/scala/org/apache/spark/sql/execution/datasources/v2/BatchScanExecShim.scala
@@ -163,18 +163,4 @@ abstract class ArrowBatchScanExecShim(original: BatchScanExec)
   override def ordering: Option[Seq[SortOrder]] = original.ordering
 
   override def output: Seq[Attribute] = original.output
-
-  override lazy val metrics = {
-    Map("numOutputRows" -> SQLMetrics.createMetric(sparkContext, "number of output rows")) ++
-      customMetrics
-  }
-
-  override def doExecuteColumnar(): RDD[ColumnarBatch] = {
-    val numOutputRows = longMetric("numOutputRows")
-    inputRDD.asInstanceOf[RDD[ColumnarBatch]].map {
-      b =>
-        numOutputRows += b.numRows()
-        b
-    }
-  }
 }
